requirements.txt
boto3>=1.34.0
python-dotenv>=1.0.0
botocore>=1.37.0
typing-extensions>=4.9.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
